Moved builder to own module, simplified Marian resource retrieval
